pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "greta-cli"
version = "0.1.0"
description = "Command-line interface for automated data analysis using genetic algorithms"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "scikit-optimize>=0.9.0",
    "deap>=1.3.0",
    "statsmodels>=0.13.0",
    "tqdm>=4.0.0",
    "fpdf>=1.7.0",
    "reportlab>=4.0.0",
    "matplotlib>=3.5.0",
    "dask>=2022.0.0",
    "dask[dataframe]>=2022.0.0",
    "streamlit>=1.28.0",
    "bcrypt>=4.0.0",
    "sqlalchemy>=2.0.0",
    "dowhy>=0.10.0",
    "pyswarms>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/your-org/greta-cli"
Repository = "https://github.com/your-org/greta-cli"
Issues = "https://github.com/your-org/greta-cli/issues"

[project.scripts]
greta = "greta_cli.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["greta_cli*", "greta_core*", "greta_web*"]

[tool.setuptools.package-data]
greta_cli = ["*.yml", "*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true